refactor: convert Analytics context to a package
